remove IndexMap old
