+++ cpp/dolfinx/common/CMakeLists.txt
@@ -3,7 +3,6 @@ set(HEADERS_common
   ${CMAKE_CURRENT_SOURCE_DIR}/dolfin_common.h
   ${CMAKE_CURRENT_SOURCE_DIR}/dolfin_doc.h
   ${CMAKE_CURRENT_SOURCE_DIR}/IndexMap.h
-  ${CMAKE_CURRENT_SOURCE_DIR}/IndexMapNew.h
   ${CMAKE_CURRENT_SOURCE_DIR}/log.h
   ${CMAKE_CURRENT_SOURCE_DIR}/loguru.hpp
   ${CMAKE_CURRENT_SOURCE_DIR}/sort.h
@@ -21,7 +20,6 @@ set(HEADERS_common
 target_sources(dolfinx PRIVATE
   ${CMAKE_CURRENT_SOURCE_DIR}/defines.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/IndexMap.cpp
-  ${CMAKE_CURRENT_SOURCE_DIR}/IndexMapNew.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/log.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/MPI.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/Table.cpp
