+++ cpp/dolfinx/common/IndexMap.h
@@ -1,4 +1,4 @@
-// Copyright (C) 2015-2019 Chris Richardson, Garth N. Wells and Igor Baratta
+// Copyright (C) 2015-2022 Chris Richardson, Garth N. Wells and Igor Baratta
 //
 // This file is part of DOLFINx (https://www.fenicsproject.org)
 //
@@ -6,14 +6,10 @@
 
 #pragma once
 
-#include <array>
+#include "IndexMap.h"
 #include <cstdint>
 #include <dolfinx/common/MPI.h>
-#include <dolfinx/graph/AdjacencyList.h>
-#include <functional>
-#include <map>
 #include <memory>
-#include <tuple>
 #include <utility>
 #include <vector>
 #include <xtl/xspan.hpp>
@@ -21,62 +17,113 @@
 namespace dolfinx::common
 {
 // Forward declaration
-class IndexMapOld;
+class IndexMap;
+
+/// @brief Given a vector of indices (local numbering, owned or ghost)
+/// and an index map, this function returns the indices owned by this
+/// process, including indices that might have been in the list of
+/// indices on another processes.
+/// @param[in] indices List of indices
+/// @param[in] map The index map
+/// @return Indices owned by the calling process
+std::vector<int32_t>
+compute_owned_indices(const xtl::span<const std::int32_t>& indices,
+                      const IndexMap& map);
+
+/// @brief Compute layout data and ghost indices for a stacked
+/// (concatenated) index map, i.e. 'splice' multiple maps into one.
+///
+/// The input maps are concatenated, with indices in `maps` and owned by
+/// the caller remaining owned by the caller. Ghost data is stored at
+/// the end of the local range as normal, with the ghosts in blocks in
+/// the order of the index maps in `maps`.
+///
+/// @note Index maps with a block size are unrolled in the data for the
+/// concatenated index map.
+/// @note Communication is required to compute the new ghost indices.
+///
+/// @param[in] maps List of (index map, block size) pairs
+/// @returns The (0) global offset of a concatenated map for the calling
+/// rank, (1) local offset for the owned indices of each submap in the
+/// concatenated map, (2) new indices for the ghosts for each submap,
+/// and (3) owner rank of each ghost entry for each submap.
+std::tuple<std::int64_t, std::vector<std::int32_t>,
+           std::vector<std::vector<std::int64_t>>,
+           std::vector<std::vector<int>>>
+stack_index_maps(
+    const std::vector<
+        std::pair<std::reference_wrapper<const common::IndexMap>, int>>& maps);
 
 /// This class represents the distribution index arrays across
 /// processes. An index array is a contiguous collection of N+1 indices
 /// [0, 1, . . ., N] that are distributed across M processes. On a given
-/// process, the IndexMapOld stores a portion of the index set using local
+/// process, the IndexMap stores a portion of the index set using local
 /// indices [0, 1, . . . , n], and a map from the local indices to a
 /// unique global index.
-class IndexMapOld
+class IndexMap
 {
 public:
-  /// Mode for reverse scatter operation
-  enum class Mode
-  {
-    insert,
-    add
-  };
-
-  /// Edge directions of neighborhood communicator
-  enum class Direction
-  {
-    reverse, // Ghost to owner
-    forward, // Owner to ghost
-  };
-
-  /// Create an index map with local_size owned indiced on this process
+  /// @brief Create an non-overlapping index map.
   ///
   /// @note Collective
+  ///
   /// @param[in] comm The MPI communicator
-  /// @param[in] local_size Local size of the IndexMapOld, i.e. the number
+  /// @param[in] local_size Local size of the index map, i.e. the number
   /// of owned entries
-  /// @param[in] dest_ranks Ranks that 'ghost' indices that are owned by
-  /// the calling rank. I.e., ranks that the caller will send data to
-  /// when updating ghost values.
+  IndexMap(MPI_Comm comm, std::int32_t local_size);
+
+  /// @brief Create an overlapping (ghosted) index map.
+  ///
+  /// @note Collective
+  ///
+  /// @param[in] comm The MPI communicator
+  /// @param[in] local_size Local size of the index map, i.e. the number
+  /// of owned entries
   /// @param[in] ghosts The global indices of ghost entries
-  /// @param[in] src_ranks Owner rank (on global communicator) of each
-  /// entry in @p ghosts
-  IndexMapOld(MPI_Comm comm, std::int32_t local_size,
-              const xtl::span<const int>& dest_ranks,
-              const xtl::span<const std::int64_t>& ghosts,
-              const xtl::span<const int>& src_ranks);
+  /// @param[in] owners Owner rank (on global communicator) of each
+  /// entry in `ghosts`
+  IndexMap(MPI_Comm comm, std::int32_t local_size,
+           const xtl::span<const std::int64_t>& ghosts,
+           const xtl::span<const int>& owners);
+
+  /// @brief Create an overlapping (ghosted) index map.
+  ///
+  /// This constructor is optimised for the case where the 'source'
+  /// (ranks that own indices ghosted by the caller) and 'destination'
+  /// ranks (ranks that ghost indices owned by the caller) are already
+  /// available. It allows the complex computation of the destination
+  /// ranks from `owners`.
+  ///
+  /// @note Collective
+  ///
+  /// @param[in] comm The MPI communicator
+  /// @param[in] local_size Local size of the index map, i.e. the number
+  /// @param[in] src_dest Lists of [0] src and [1] dest ranks. The list
+  /// in each must be sorted and contain duplicates. `src` ranks are
+  /// owners of the indices in `ghosts`. `dest` ranks are the rank that
+  /// ghost indices owned by the caller.
+  /// @param[in] ghosts The global indices of ghost entries
+  /// @param[in] owners Owner rank (on global communicator) of each entry
+  /// in `ghosts`
+  IndexMap(MPI_Comm comm, std::int32_t local_size,
+           const std::array<std::vector<int>, 2>& src_dest,
+           const xtl::span<const std::int64_t>& ghosts,
+           const xtl::span<const int>& owners);
 
   // Copy constructor
-  IndexMapOld(const IndexMapOld& map) = delete;
+  IndexMap(const IndexMap& map) = delete;
 
   /// Move constructor
-  IndexMapOld(IndexMapOld&& map) = default;
+  IndexMap(IndexMap&& map) = default;
 
   /// Destructor
-  ~IndexMapOld() = default;
+  ~IndexMap() = default;
 
   /// Move assignment
-  IndexMapOld& operator=(IndexMapOld&& map) = default;
+  IndexMap& operator=(IndexMap&& map) = default;
 
   // Copy assignment
-  IndexMapOld& operator=(const IndexMapOld& map) = delete;
+  IndexMap& operator=(const IndexMap& map) = delete;
 
   /// Range of indices (global) owned by this process
   std::array<std::int64_t, 2> local_range() const noexcept;
@@ -94,296 +141,88 @@ public:
   /// range)
   const std::vector<std::int64_t>& ghosts() const noexcept;
 
-  /// Return the MPI communicator used to create the index map
+  /// @brief Return the MPI communicator that the map is defined on.
   /// @return Communicator
   MPI_Comm comm() const;
 
-  /// Return a MPI communicator with attached distributed graph topology
-  /// information
-  /// @param[in] dir Edge direction of communicator (forward, reverse)
-  /// @return A neighborhood communicator for the specified edge direction
-  MPI_Comm comm(Direction dir) const;
+  /// @brief Compute global indices for array of local indices.
+  /// @param[in] local Local indices
+  /// @param[out] global The global indices
+  void local_to_global(const xtl::span<const std::int32_t>& local,
+                       const xtl::span<std::int64_t>& global) const;
 
-  /// Local (owned) indices shared with neighbor processes, i.e. are
-  /// ghosts on other processes, grouped by sharing (neighbor) process
-  /// (destination ranks in forward communicator and source ranks in the
-  /// reverse communicator). `scatter_fwd_indices().links(p)` gives the
-  /// list of owned indices that needs to be sent to neighbourhood rank
-  /// `p` during a forward scatter.
+  /// @brief Compute local indices for array of global indices
+  /// @param[in] global Global indices
+  /// @param[out] local The local of the corresponding global index in
+  /// 'global'. Returns -1 if the local index does not exist on this
+  /// process.
+  void global_to_local(const xtl::span<const std::int64_t>& global,
+                       const xtl::span<std::int32_t>& local) const;
+
+  /// @brief Build list of indices with global indexing.
+  /// @return The global index for all local indices (0, 1, 2, ...) on
+  /// this process, including ghosts
+  std::vector<std::int64_t> global_indices() const;
+
+  /// @brief The ranks that own each ghost index.
+  /// @return List of ghost owners. The owning rank of the ith ghost
+  /// index is `owners()[i]`.
+  const std::vector<int>& owners() const { return _owners; }
+
+  /// @brief Create new index map from a subset of indices in this index
+  /// map.
   ///
-  /// Entries are ordered such that `scatter_fwd_indices.offsets()` is
-  /// the send displacement array for a forward scatter and
-  /// `scatter_fwd_indices.array()[i]` in the index of the owned index
-  /// that should be placed at position `i` in the send buffer for a
-  /// forward scatter.
-  /// @return List of indices that are ghosted on other processes
-  const graph::AdjacencyList<std::int32_t>& scatter_fwd_indices() const;
-
-  /// Position of ghost entries in the receive buffer after a forward
-  /// scatter, e.g. for a receive buffer `b` and a set operation, the
-  /// ghost values should be updated  by `ghost_value[i] =
-  /// b[scatter_fwd_ghost_positions[i]]`.
-  /// @return Position of the ith ghost entry in the received buffer
-  const std::vector<std::int32_t>& scatter_fwd_ghost_positions() const noexcept;
-
-  /// @brief Compute the owner on the neighborhood communicator of each
-  /// ghost index.
+  /// The order of the owned indices is preserved, with new map
+  /// effectively a 'compressed' map.
   ///
-  /// The neighborhood ranks are the 'source' ranks on the 'reverse'
-  /// communicator, i.e. the neighborhood source ranks on the
-  /// communicator returned by
-  /// IndexMapOld::comm(IndexMapOld::Direction::reverse). The source ranks on
-  /// IndexMapOld::comm(IndexMapOld::Direction::reverse) communicator can be
-  /// used to convert the returned neighbour ranks to the rank indices on
-  /// the full communicator.
+  /// @param[in] indices Local indices in the map that should appear in
+  /// the new index map. All indices must be owned, i.e. indices must be
+  /// less than `this->size_local()`.
+  /// @pre `indices` must be sorted and contain no duplicates.
+  /// @return The (i) new index map and (ii) a map from the ghost
+  /// position in the new map to the ghost position in the original
+  /// (this) map
+  std::pair<IndexMap, std::vector<std::int32_t>>
+  create_submap(const xtl::span<const std::int32_t>& indices) const;
+
+  /// @todo Aim to remove this function?
   ///
-  /// @return The owning rank on the neighborhood communicator of the
-  /// ith ghost index.
-  std::vector<int> ghost_owners() const;
+  /// @brief Compute map from each local (owned) index to the set of
+  /// ranks that have the index as a ghost.
+  /// @return shared indices
+  graph::AdjacencyList<int> index_to_dest_ranks() const;
 
-  /// Start a non-blocking send of owned data to ranks that ghost the
-  /// data. The communication is completed by calling
-  /// IndexMapOld::scatter_fwd_end. The send and receive buffer should not
-  /// be changed until after IndexMapOld::scatter_fwd_end has been called.
+  /// @brief Build a list of owned indices that are ghosted by another
+  /// rank.
+  /// @return The local index of owned indices that are ghosts on other
+  /// rank(s). The indicies are unique and sorted.
+  std::vector<std::int32_t> shared_indices() const;
+
+  /// @brief Ordered set of MPI ranks that own caller's ghost indices.
   ///
-  /// @param[in] send_buffer Local data associated with each owned local
-  /// index to be sent to process where the data is ghosted. It must not
-  /// be changed until after a call to IndexMapOld::scatter_fwd_end. The
-  /// order of data in the buffer is given by
-  /// IndexMapOld::scatter_fwd_indices.
-  /// @param data_type The MPI data type. To send data with a block size
-  /// use `MPI_Type_contiguous` with size `n`
-  /// @param request The MPI request handle for tracking the status of
-  /// the non-blocking communication
-  /// @param recv_buffer A buffer used for the received data. The
-  /// position of ghost entries in the buffer is given by
-  /// IndexMapOld::scatter_fwd_ghost_positions. The buffer must not be
-  /// accessed or changed until after a call to
-  /// IndexMapOld::scatter_fwd_end.
-  template <typename T>
-  void scatter_fwd_begin(const xtl::span<const T>& send_buffer,
-                         MPI_Datatype& data_type, MPI_Request& request,
-                         const xtl::span<T>& recv_buffer) const
-  {
-    // Send displacement
-    const std::vector<int32_t>& displs_send_fwd = _shared_indices->offsets();
-
-    // Return early if there are no incoming or outgoing edges
-    if (_displs_recv_fwd.size() == 1 and displs_send_fwd.size() == 1)
-      return;
-
-    // Get block size
-    int n;
-    MPI_Type_size(data_type, &n);
-    n /= sizeof(T);
-    if (static_cast<int>(send_buffer.size()) != n * displs_send_fwd.back())
-      throw std::runtime_error("Incompatible send buffer size.");
-    if (static_cast<int>(recv_buffer.size()) != n * _displs_recv_fwd.back())
-      throw std::runtime_error("Incompatible receive buffer size..");
-
-    // Start send/receive
-    MPI_Ineighbor_alltoallv(send_buffer.data(), _sizes_send_fwd.data(),
-                            displs_send_fwd.data(), data_type,
-                            recv_buffer.data(), _sizes_recv_fwd.data(),
-                            _displs_recv_fwd.data(), data_type,
-                            _comm_owner_to_ghost.comm(), &request);
-  }
-
-  /// Complete a non-blocking send from the local owner of to process
-  /// ranks that have the index as a ghost. This function complete the
-  /// communication started by IndexMapOld::scatter_fwd_begin.
+  /// Typically used when creating neighbourhood communicators.
   ///
-  /// @param[in] request The MPI request handle for tracking the status
-  /// of the send
-  void scatter_fwd_end(MPI_Request& request) const
-  {
-    // Return early if there are no incoming or outgoing edges
-    const std::vector<int32_t>& displs_send_fwd = _shared_indices->offsets();
-    if (_displs_recv_fwd.size() == 1 and displs_send_fwd.size() == 1)
-      return;
+  /// @return MPI ranks than own ghost indices.  The ranks are unique
+  /// and sorted.
+  const std::vector<int>& src() const noexcept;
 
-    // Wait for communication to complete
-    MPI_Wait(&request, MPI_STATUS_IGNORE);
-  }
-
-  /// Send n values for each index that is owned to processes that have
-  /// the index as a ghost. The size of the input array local_data must
-  /// be the same as n * size_local().
+  /// @brief Ordered set of MPI ranks that ghost indices owned by
+  /// caller.
   ///
-  /// @param[in] local_data Local data associated with each owned local
-  /// index to be sent to process where the data is ghosted. Size must
-  /// be n * size_local().
-  /// @param[in,out] remote_data Ghost data on this process received
-  /// from the owning process. Size will be n * num_ghosts().
-  /// @param[in] n Number of data items per index
-  template <typename T>
-  void scatter_fwd(const xtl::span<const T>& local_data,
-                   xtl::span<T> remote_data, int n) const
-  {
-    MPI_Datatype data_type;
-    if (n == 1)
-      data_type = dolfinx::MPI::mpi_type<T>();
-    else
-    {
-      MPI_Type_contiguous(n, dolfinx::MPI::mpi_type<T>(), &data_type);
-      MPI_Type_commit(&data_type);
-    }
-
-    const std::vector<std::int32_t>& indices = _shared_indices->array();
-    std::vector<T> send_buffer(n * indices.size());
-    for (std::size_t i = 0; i < indices.size(); ++i)
-    {
-      std::copy_n(std::next(local_data.cbegin(), n * indices[i]), n,
-                  std::next(send_buffer.begin(), n * i));
-    }
-
-    MPI_Request request;
-    std::vector<T> buffer_recv(n * _displs_recv_fwd.back());
-    scatter_fwd_begin(xtl::span<const T>(send_buffer), data_type, request,
-                      xtl::span<T>(buffer_recv));
-    scatter_fwd_end(request);
-
-    // Copy into ghost area ("remote_data")
-    assert(remote_data.size() == n * _ghost_pos_recv_fwd.size());
-    for (std::size_t i = 0; i < _ghost_pos_recv_fwd.size(); ++i)
-    {
-      std::copy_n(std::next(buffer_recv.cbegin(), n * _ghost_pos_recv_fwd[i]),
-                  n, std::next(remote_data.begin(), n * i));
-    }
-
-    if (n != 1)
-      MPI_Type_free(&data_type);
-  }
-
-  /// Start a non-blocking send of ghost values to the owning rank. The
-  /// non-blocking communication is completed by calling
-  /// IndexMapOld::scatter_rev_end. A reverse scatter is the transpose of
-  /// IndexMapOld::scatter_fwd_begin.
+  /// Typically used when creating neighbourhood communicators.
   ///
-  /// @param[in] send_buffer Send buffer filled with ghost data on this
-  /// process to be sent to the owning rank. The order of the data is
-  /// given by IndexMapOld::scatter_fwd_ghost_positions, with
-  /// IndexMapOld::scatter_fwd_ghost_positions()[i] being the index of the
-  /// ghost data that should be placed in position `i` of the buffer.
-  /// @param data_type The MPI data type. To send data with a block size
-  /// use `MPI_Type_contiguous` with size `n`
-  /// @param request The MPI request handle for tracking the status of
-  /// the send
-  /// @param recv_buffer A buffer used for the received data. It must
-  /// not be changed until after a call to IndexMapOld::scatter_rev_end.
-  /// The ordering of the data is given by
-  /// IndexMapOld::scatter_fwd_indices, with
-  /// IndexMapOld::scatter_fwd_indices()[i] being the position in the owned
-  /// data array that corresponds to position `i` in the buffer.
-  template <typename T>
-  void scatter_rev_begin(const xtl::span<const T>& send_buffer,
-                         MPI_Datatype& data_type, MPI_Request& request,
-                         const xtl::span<T>& recv_buffer) const
-  {
-    // Get displacement vector
-    const std::vector<int32_t>& displs_send_fwd = _shared_indices->offsets();
+  /// @return MPI ranks than own ghost indices. The ranks are unique
+  /// and sorted.
+  const std::vector<int>& dest() const noexcept;
 
-    // Return early if there are no incoming or outgoing edges
-    if (_displs_recv_fwd.size() == 1 and displs_send_fwd.size() == 1)
-      return;
-
-    // Get block size
-    int n;
-    MPI_Type_size(data_type, &n);
-    n /= sizeof(T);
-    if (static_cast<int>(send_buffer.size()) != n * _ghosts.size())
-      throw std::runtime_error("Inconsistent send buffer size.");
-    if (static_cast<int>(recv_buffer.size()) != n * displs_send_fwd.back())
-      throw std::runtime_error("Inconsistent receive buffer size.");
-
-    // Send and receive data
-    MPI_Ineighbor_alltoallv(send_buffer.data(), _sizes_recv_fwd.data(),
-                            _displs_recv_fwd.data(), data_type,
-                            recv_buffer.data(), _sizes_send_fwd.data(),
-                            displs_send_fwd.data(), data_type,
-                            _comm_ghost_to_owner.comm(), &request);
-  }
-
-  /// Complete a non-blocking send of ghost values to the owning rank.
-  /// This function complete the communication started by
-  /// IndexMapOld::scatter_rev_begin.
+  /// @brief Check if index map has overlaps (ghosts on any rank).
   ///
-  /// @param[in] request The MPI request handle for tracking the status
-  /// of the send
-  void scatter_rev_end(MPI_Request& request) const
-  {
-    // Return early if there are no incoming or outgoing edges
-    const std::vector<int32_t>& displs_send_fwd = _shared_indices->offsets();
-    if (_displs_recv_fwd.size() == 1 and displs_send_fwd.size() == 1)
-      return;
-
-    // Wait for communication to complete
-    MPI_Wait(&request, MPI_STATUS_IGNORE);
-  }
-
-  /// Send n values for each ghost index to owning to the process
+  /// The return value of this function is determined by which
+  /// constructor was used to create the index map.
   ///
-  /// @param[in,out] local_data Local data associated with each owned
-  /// local index to be sent to process where the data is ghosted. Size
-  /// must be n * size_local().
-  /// @param[in] remote_data Ghost data on this process received from
-  /// the owning process. Size will be n * num_ghosts().
-  /// @param[in] n Number of data items per index
-  /// @param[in] op Sum or set received values in local_data
-  template <typename T>
-  void scatter_rev(xtl::span<T> local_data,
-                   const xtl::span<const T>& remote_data, int n,
-                   IndexMapOld::Mode op) const
-  {
-    MPI_Datatype data_type;
-    if (n == 1)
-      data_type = dolfinx::MPI::mpi_type<T>();
-    else
-    {
-      MPI_Type_contiguous(n, dolfinx::MPI::mpi_type<T>(), &data_type);
-      MPI_Type_commit(&data_type);
-    }
-
-    // Pack send buffer
-    std::vector<T> buffer_send;
-    buffer_send.resize(n * _displs_recv_fwd.back());
-    for (std::size_t i = 0; i < _ghost_pos_recv_fwd.size(); ++i)
-    {
-      std::copy_n(std::next(remote_data.cbegin(), n * i), n,
-                  std::next(buffer_send.begin(), n * _ghost_pos_recv_fwd[i]));
-    }
-
-    // Exchange data
-    MPI_Request request;
-    std::vector<T> buffer_recv(n * _shared_indices->array().size());
-    scatter_rev_begin(xtl::span<const T>(buffer_send), data_type, request,
-                      xtl::span<T>(buffer_recv));
-    scatter_rev_end(request);
-
-    // Copy or accumulate into "local_data"
-    assert(local_data.size() == n * this->size_local());
-    const std::vector<std::int32_t>& shared_indices = _shared_indices->array();
-    switch (op)
-    {
-    case Mode::insert:
-      for (std::size_t i = 0; i < shared_indices.size(); ++i)
-      {
-        std::copy_n(std::next(buffer_recv.cbegin(), n * i), n,
-                    std::next(local_data.begin(), n * shared_indices[i]));
-      }
-      break;
-    case Mode::add:
-      for (std::size_t i = 0; i < shared_indices.size(); ++i)
-      {
-        for (int j = 0; j < n; ++j)
-          local_data[shared_indices[i] * n + j] += buffer_recv[i * n + j];
-      }
-      break;
-    }
-
-    if (n != 1)
-      MPI_Type_free(&data_type);
-  }
+  /// @return True if index map has overlaps on any ranks, otherwise
+  /// false.
+  bool overlapped() const noexcept;
 
 private:
   // Range of indices (global) owned by this process
@@ -392,44 +231,22 @@ private:
   // Number indices across communicator
   std::int64_t _size_global;
 
-  // MPI communicator (duplicated of 'input' communicator)
+  // MPI communicator that map is defined on
   dolfinx::MPI::Comm _comm;
-
-  // Communicator where the source ranks own the indices in the callers
-  // halo, and the destination ranks 'ghost' indices owned by the
-  // caller. I.e.,
-  // - in-edges (src) are from ranks that own my ghosts
-  // - out-edges (dest) go to ranks that 'ghost' my owned indices
-  dolfinx::MPI::Comm _comm_owner_to_ghost;
-
-  // Communicator where the source ranks have ghost indices that are
-  // owned by the caller, and the destination ranks are the owners of
-  // indices in the callers halo region. I.e.,
-  // - in-edges (src) are from ranks that 'ghost' my owned indicies
-  // - out-edges (dest) are to the owning ranks of my ghost indices
-  dolfinx::MPI::Comm _comm_ghost_to_owner;
-
-  // MPI sizes and displacements for forward (owner -> ghost) scatter
-  // Note: '_displs_send_fwd' can be got from _shared_indices->offsets()
-  std::vector<std::int32_t> _sizes_send_fwd, _sizes_recv_fwd, _displs_recv_fwd;
-
-  // Position in the recv buffer for a forward scatter for the ith ghost
-  // index (_ghost[i]) entry
-  std::vector<std::int32_t> _ghost_pos_recv_fwd;
 
   // Local-to-global map for ghost indices
   std::vector<std::int64_t> _ghosts;
 
-  // Ghost owners (global ranks)
+  // Local-to-global map for ghost indices
   std::vector<int> _owners;
 
-  // List of owned local indices that are in the ghost (halo) region on
-  // other ranks, grouped by rank in the neighbor communicator
-  // (destination ranks in forward communicator and source ranks in the
-  // reverse communicator), i.e. `_shared_indices.num_nodes() ==
-  // size(_comm_owner_to_ghost)`. The array _shared_indices.offsets() is
-  // equivalent to 'displs_send_fwd'.
-  std::unique_ptr<graph::AdjacencyList<std::int32_t>> _shared_indices;
-};
+  // Set of ranks that own ghosts
+  std::vector<int> _src;
 
+  // Set of ranks ghost owned indices
+  std::vector<int> _dest;
+
+  // True if map has overlaps (ghosts)
+  bool _overlapping;
+};
 } // namespace dolfinx::common
