+++ cpp/dolfinx/common/Scatterer.h
@@ -6,7 +6,7 @@
 
 #pragma once
 
-#include <dolfinx/common/IndexMapNew.h>
+#include <dolfinx/common/IndexMap.h>
 #include <dolfinx/common/sort.h>
 #include <dolfinx/graph/AdjacencyList.h>
 #include <memory>
