+++ cpp/dolfinx/fem/DirichletBC.cpp
@@ -9,7 +9,7 @@
 #include "FiniteElement.h"
 #include <algorithm>
 #include <array>
-#include <dolfinx/common/IndexMapNew.h>
+#include <dolfinx/common/IndexMap.h>
 #include <dolfinx/common/sort.h>
 #include <dolfinx/graph/AdjacencyList.h>
 #include <dolfinx/mesh/Mesh.h>
