+++ cpp/dolfinx/fem/assemble_scalar_impl.h
@@ -10,7 +10,7 @@
 #include "Form.h"
 #include "FunctionSpace.h"
 #include "utils.h"
-#include <dolfinx/common/IndexMapNew.h>
+#include <dolfinx/common/IndexMap.h>
 #include <dolfinx/common/utils.h>
 #include <dolfinx/mesh/Geometry.h>
 #include <dolfinx/mesh/Mesh.h>
