+++ cpp/dolfinx/fem/assemble_vector_impl.h
@@ -12,7 +12,7 @@
 #include "Form.h"
 #include "FunctionSpace.h"
 #include "utils.h"
-#include <dolfinx/common/IndexMapNew.h>
+#include <dolfinx/common/IndexMap.h>
 #include <dolfinx/common/utils.h>
 #include <dolfinx/graph/AdjacencyList.h>
 #include <dolfinx/mesh/Geometry.h>
