+++ cpp/dolfinx/fem/sparsitybuild.cpp
@@ -6,7 +6,7 @@
 
 #include "sparsitybuild.h"
 #include "DofMap.h"
-#include <dolfinx/common/IndexMapNew.h>
+#include <dolfinx/common/IndexMap.h>
 #include <dolfinx/graph/AdjacencyList.h>
 #include <dolfinx/la/SparsityPattern.h>
 #include <dolfinx/mesh/Topology.h>
