+++ cpp/dolfinx/fem/utils.cpp
@@ -14,7 +14,7 @@
 #include "dofmapbuilder.h"
 #include "sparsitybuild.h"
 #include <array>
-#include <dolfinx/common/IndexMapNew.h>
+#include <dolfinx/common/IndexMap.h>
 #include <dolfinx/common/Timer.h>
 #include <dolfinx/common/log.h>
 #include <dolfinx/la/SparsityPattern.h>
