+++ cpp/dolfinx/geometry/BoundingBoxTree.cpp
@@ -7,7 +7,7 @@
 
 #include "BoundingBoxTree.h"
 #include "utils.h"
-#include <dolfinx/common/IndexMapNew.h>
+#include <dolfinx/common/IndexMap.h>
 #include <dolfinx/common/log.h>
 #include <dolfinx/common/utils.h>
 #include <dolfinx/mesh/Geometry.h>
