+++ cpp/dolfinx/io/vtk_utils.cpp
@@ -6,7 +6,7 @@
 
 #include "vtk_utils.h"
 #include "cells.h"
-#include <dolfinx/common/IndexMapNew.h>
+#include <dolfinx/common/IndexMap.h>
 #include <dolfinx/fem/DofMap.h>
 #include <dolfinx/fem/FiniteElement.h>
 #include <dolfinx/fem/FunctionSpace.h>
