+++ cpp/dolfinx/io/xdmf_mesh.cpp
@@ -9,7 +9,7 @@
 #include "pugixml.hpp"
 #include "xdmf_read.h"
 #include "xdmf_utils.h"
-#include <dolfinx/common/IndexMapNew.h>
+#include <dolfinx/common/IndexMap.h>
 #include <dolfinx/fem/ElementDofLayout.h>
 
 using namespace dolfinx;
