+++ cpp/dolfinx/io/xdmf_meshtags.h
@@ -9,7 +9,7 @@
 #include "pugixml.hpp"
 #include "xdmf_mesh.h"
 #include "xdmf_utils.h"
-#include <dolfinx/common/IndexMapNew.h>
+#include <dolfinx/common/IndexMap.h>
 #include <dolfinx/common/MPI.h>
 #include <dolfinx/common/log.h>
 #include <dolfinx/mesh/MeshTags.h>
