+++ cpp/dolfinx/la/MatrixCSR.h
@@ -8,7 +8,7 @@
 
 #include "SparsityPattern.h"
 #include <dolfinx/common/IndexMap.h>
-#include <dolfinx/common/IndexMapNew.h>
+#include <dolfinx/common/IndexMap.h>
 #include <dolfinx/common/MPI.h>
 #include <dolfinx/graph/AdjacencyList.h>
 #include <mpi.h>
