+++ cpp/dolfinx/la/SparsityPattern.cpp
@@ -6,7 +6,7 @@
 
 #include "SparsityPattern.h"
 #include <algorithm>
-#include <dolfinx/common/IndexMapNew.h>
+#include <dolfinx/common/IndexMap.h>
 #include <dolfinx/common/MPI.h>
 #include <dolfinx/common/Timer.h>
 #include <dolfinx/common/log.h>
