+++ cpp/dolfinx/mesh/Mesh.cpp
@@ -12,7 +12,7 @@
 #include "graphbuild.h"
 #include "topologycomputation.h"
 #include "utils.h"
-#include <dolfinx/common/IndexMapNew.h>
+#include <dolfinx/common/IndexMap.h>
 #include <dolfinx/common/utils.h>
 #include <dolfinx/fem/CoordinateElement.h>
 #include <dolfinx/graph/AdjacencyList.h>
