+++ cpp/dolfinx/mesh/MeshTags.h
@@ -10,7 +10,7 @@
 #include "Mesh.h"
 #include "Topology.h"
 #include <algorithm>
-#include <dolfinx/common/IndexMapNew.h>
+#include <dolfinx/common/IndexMap.h>
 #include <dolfinx/common/log.h>
 #include <dolfinx/common/utils.h>
 #include <dolfinx/graph/AdjacencyList.h>
