+++ cpp/dolfinx/mesh/utils.cpp
@@ -11,7 +11,7 @@
 #include "graphbuild.h"
 #include <algorithm>
 #include <cstdlib>
-#include <dolfinx/common/IndexMapNew.h>
+#include <dolfinx/common/IndexMap.h>
 #include <dolfinx/common/log.h>
 #include <dolfinx/common/math.h>
 #include <dolfinx/fem/ElementDofLayout.h>
