+++ cpp/dolfinx/refinement/plaza.cpp
@@ -6,7 +6,7 @@
 
 #include "plaza.h"
 #include "utils.h"
-#include <dolfinx/common/IndexMapNew.h>
+#include <dolfinx/common/IndexMap.h>
 #include <dolfinx/common/Timer.h>
 #include <dolfinx/graph/AdjacencyList.h>
 #include <dolfinx/mesh/Geometry.h>
