+++ cpp/test/matrix.cpp
@@ -9,8 +9,8 @@
 #include "poisson.h"
 #include <catch2/catch.hpp>
 #include <dolfinx.h>
-#include <dolfinx/common/IndexMapNew.h>
-#include <dolfinx/common/IndexMapNew.h>
+#include <dolfinx/common/IndexMap.h>
+#include <dolfinx/common/IndexMap.h>
 #include <dolfinx/la/MatrixCSR.h>
 #include <dolfinx/la/SparsityPattern.h>
 #include <dolfinx/la/Vector.h>
