+++ cpp/test/vector.cpp
@@ -8,7 +8,7 @@
 
 #include <catch2/catch.hpp>
 #include <dolfinx.h>
-#include <dolfinx/common/IndexMapNew.h>
+#include <dolfinx/common/IndexMap.h>
 #include <dolfinx/la/MatrixCSR.h>
 #include <dolfinx/la/SparsityPattern.h>
 #include <dolfinx/la/Vector.h>
